nodes/4.1/4.1.13: Fix errata link from 45273 -> 2547
The broken ID snuck into this release.  Unfortunately, it's baked into
the release image too:

  $ oc adm release info quay.io/openshift-release-dev/ocp-release:4.1.13 | grep errata
      url: https://access.redhat.com/errata/RHBA-2019:45273

and while we can't fix that without getting a new release digest and
needing a new signature, we can update the metadata in this repo.

We have an internal ticket to get a 403 from the broken 45273 to the
intended 2547 [1], and if that happens we might revert this commit.

[1]: https://projects.engineering.redhat.com/browse/CPPP-4572