hack/graph-util: Pull nodes and metadata from Quay
Each time before updating Quay.  This is slow (tens of seconds), but
means release admins don't have to bother tracking the nodes and their
default edges in this repository (personally, I'd rather track nodes
and edges [1], but I'm not a release admin ;).  The delay is mitigated
by keeping a local cache (outside of version control) of the node
metadata.

[1]: https://github.com/openshift/cincinnati-graph-data/pull/1