hack/graph-util: Fix layer iteration
If the top layer is missing the metadata file, catch the error and
continue to look at future layers.  For example, [1] has the metadata
file in an earlier layer.

This means we also have to fix the not-found-in-any-layer error to use
the payload instead of node['version'], because load_nodes can call
get_release_metadata on:

  node = {'payload': pullspec}

and the 'version' string isn't present (we were trying to extract it
from the metadata file, which we didn't find).

[1]: quay.io/openshift-release-dev/ocp-release:4.0.0-0.9