channels/candidate-4.3: Promote 4.2.16+amd64
792cc4c3d7 (Enable the 4.2.16 to 4.3.0 upgrade, 2020-02-03, #40)
pulled the block to allow this edge, but to get an edge in a given
channel, both the source and target release need to be in that
channel.  This was added to fast-4.3 in 792cc4c3d7 and later
restricted to amd64 in 9a25271ad5 (Do not add 4.2 s390 release to 4.3
channels, 2020-02-03, #41), but we'd skipped over candidate-4.3.