channels: Add 4.2.16+amd64 back to candidate-4.3 and fast-4.3
Once a node enters a channel, we don't want to remove it.  Doing so
would leave existing residents in that channel with the
VersionNotFound error out of the cluster-version operator.  Restore
4.2.16+amd64 to the channels it was in before c22d57ef1b
(blocked-edges: Block edges for storage.conf machine-config bug,
2020-02-19, #59), and rely on blocked-edges alone (also from
c22d57ef1b) to keep 4.2.16 -> 4.3.0 out of the graph.

Folks on 4.2.16 in candidate-4.3 or fast-4.3 will now be suggested to
take the 4.2.16 -> 4.2.18 -> 4.3.1 pathway into 4.3.