hack/stabilization-changes: Use 'git blame' to recommend auto-stabilization
Picking the versions out of the blame data by using a '- ' prefix is a
bit of a hack, but it gets the job done for now.

Also suggest wording for the channel-bumping commit message, like:

  channels/fast-4: Promote 4.1.18 to fast-4.2

  It was promoted the feeder candidate-4.2 by 2c088bc147 (channels:
  Convert to Eric's proposed channel format, 2019-12-05)

This is in a separate script from graph-util.py to make it more
obvious that stabilization recommendations (which suggest changes to
channel membership) are distinct from graph-util.py's existing
push-to-quay logic (which is about transmitting the channel/edge
configuration from this repository to Cincinnati).