channels: Promote 4.4.0-rc.1 after all
Blocking edges into it, as discussed in d544dde953
(blocked-edges/4.2.23: Block all incoming edges on the service CA bug
1810036, 2020-03-18, #127).  But adding the release itself into the
channel for fresh installs, because blowing up after 13 months when
the buggy CA rotation fires is unlikley to impact customers (who will
probably not run an RC for that long).  I was initially expecting us
to just cut a fresh RC with the fix, but it turns out that cutting RCs
is non-trivial and we only expect them weekly.  The 13-month issue is
not important enough to make folks wait a week for a new RC.