hack/graph-util: Drop --token
Now that Cincinnati is reading directly from the master branch, there
is no need to manually publish to Quay labels.  Drop the --token
argument to help make that clear.  We can't drop this script
completely yet, because we need to slot in some Rust binary to cover
pre-merge CI (e.g. "channels/whatever.yaml talks about 4.5.6.  Is
there at least one release on quay.io with that version baked in?").