channels/fast-*: Bump delay to one week
We usually cut candidates on Wednesday, cook on Thursday and Friday,
and then promote into fast on Monday.  And the real gate for fast is
the errata going public, which can be tracked with 282cf71b11
(hack/errata: Poll for recent errata events, 2020-04-02, #156).  Tune
the delay here up to a week so the script doesn't recommend fast
updates while we are waiting for the public errata.

An alternative would be to drop delay from the fast channels entirely
and gate purely on production errata.  I'm not quite ready for that
yet, and want the delay as a backstop for forgotten errata.

Also bump the stable-4.1 delay to one week, because there is no
fast-4.1, so stable-4.1 should also be triggered by errata
publication.