blocked-edges/4.4.3: Block 4.3 -> 4.4.3 on etcd DNS SAN peer certs
Describeed in [1].  The rate of hitting this bug seemed low, and in
most cases clusters recovered automatically without admin
intervention, so we did not pull the edge while 4.3 -> 4.4.3 was in
fast-4.4.  But now that there's a supported 4.3 -> 4.4.4 route that is
not exposed to the bug [2], we want to block the 4.3 -> 4.4.3 edges to
remove even the small chance of getting stuck.

Also in this space is updates getting stuck on
NodeInstaller_InstallerPodFailed.  But there's a reasonably easy
workaround [3] for folks who do get stuck, so not worth pulling the
edge on that account.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1830510#c10
[2]: https://github.com/openshift/cincinnati-graph-data/pull/251
[3]: https://bugzilla.redhat.com/show_bug.cgi?id=1830409#c7