channels/*-4.5: Add 4.4.6 through 4.4.10 to fast-4.5 and stable-4.5
4.4.6 has been in stable-4.4 since dc06ce1fe3 (Enable 4.4.6 in stable
channel(s), 2020-05-27, #263) landed.  Promoting into fast/stable
channels for 4.5 makes it clear that we support 4.4.6 regardless of
whether a cluster is interested in 4.5 update recommendations or not.

And also promote 4.4.8, 4.4.9, and 4.4.10 to catch up with 61e8f69812
(Enable 4.4.8 in stable channel(s), 2020-06-10, #280), 64506bb211
(Enable 4.4.9 in stable channel(s), 2020-06-18, #289), and 92080fede0
(Enable 4.4.10 in stable channel(s), 2020-06-24, #296).