errata.py: autoLGTM the fast PR after errata has shipped
This change would add a comment with the LGTM to the PR promoting to
fast after errata has changed status to SHIPPED_LIVE. After datagrepper
gets notified that errata state has changed, it would parse PRs
targetting fast (expects default title), extract errata number and
compare to the received message. If errata numbers match the script
would comment with `/lgtm`.

This requires the hook to be deployed with GITHUB_TOKEN set, which
would be taken care of in internal autoreport repo