Add build time Upgrades suggestions
This adds minor_min, minor_block_list, minor_max, z_min, z_block_list,
and z_max which define the minimum, a list of blocked versions, and a
maximum version which can be used to filter upgrade edges. It's expected
that this is applied atop a list of versions already present in a
channel so the block_list may not see much use.

There's a default as well as arch specific definitions. If you define
arch specific data you are expected to define the entire structure.