hack/backfill: Add a backfilling script
Following up on 646535a052 (Add all previous minor stable versions to
current minor fast and stable, 2020-10-29, #526) with improved
backfilling logic.  The idea is that all previous-minor releases in
the previous minor's branch of the same weight are in your branch
(e.g. 4.3.0 is in candidate-4.3, so it gets backfilled into
candidate-4.4).  And all current-minor releases in the next-minor's
branch of the same weight are in your branch (e.g. if 4.5.0 was in
candidate-4.6, it would be backfilled into candidate-4.5, but
currently no changes from this logic).