channels: More backfilling (e.g. 4.3.0 in candidate-4.4)
Generated with:

  $ hack/backfill.py
  $ git checkout HEAD -- channels/candidate-4.3.yaml
  $ emacs channels/candidate-4.4.yaml
  $ git checkout HEAD -- channels/stable-4.6.yaml

With the candidate-4.3 and 4.4 massaging so we don't get:

   - 4.3.14
  +- 4.3.14+amd64

and similar, where all arches went into candidate-4.3, but only amd64
went into the candidate-4.4, 5ff491a42f (Enable 4.3.14 in candidate
channel(s), 2020-04-22, #193).

The stable-4.6 massaging is because we are delaying the 4.5->4.6
stable updates while we convince ourselves that fast traffic is
sufficiently reliable.