hack/errata: Log errata whose synopsis doesn't match our regular expression
We've had a checkered history with these, including:

* 4f969bfd43 (hack/errata: Match 'OpenShift Container Platform 4.6 GA
  Images' synopsis, 2020-10-27, #522).
* d161733812 (hack/errata: include impact, 2020-10-22, #513) to
  include 'Low: OpenShift Container Platform 4.3.40 security and bug
  fix update'.
* ee59956db5 (hack/errata: Extract version from errata synopsis,
  2020-09-30, #480) to exclude 'OpenShift Container Platform 4.5.11
  optional CSI driver Operators bug fix update.

Log the excluded synopses, so it's easier to find errata which we are
incorrectly excluding (by looking in the errata.py logs, instead of
hitting DataGrepper directly and then applying the synopsis matching
by hand).  Add an additional cache so we don't log our exclusion of a
given synopsis in successive polls.  Don't store the new cache on
disk, because logging a synopsis once again after a container restart
is not a big deal.