blocked-edges/4.6.6: Block * -> 4.6.(6|7) on born-in-4.1 ingress
There aren't all that many born-in-4.1 clusters within striking
distance of 4.6.6+, so the number of vulnerable clusters is small.
But *.apps breaking would be really bad, so block the edges to protect
that small subset.

We could keep 4.6.6 -> 4.6.7 in place, but I didn't bother to figure
out the regexp for "all versions that aren't 4.6.6".