graph-data.rs: verify channel order
This adds a CI tests which ensures, that releases are properly ordered: first it must be placed in candidate, then fast and then stable. If a release is put in candidate and then in stable, it would throw an error.

EUS channel is omitted for now, as its not yet entirely clear how releases would be structured there.