channels: Drop 'delay' from fast promotions
Now that we can reliably detect the public errata, that's when we want
to trigger fast promotions.  I'd expected us to want a backstop at a
week for promoting if we had trouble shipping the errata for some
reason, but after internal discussion, the consensus is that we want
the robots to point the delay out to us, but not act until the release
becomes public.  With this commit, we'll still get the:

  waiting to promote {version} to {channel}...

notifications reminding us that the release has not been tombstoned,
but that the errata blocking the promotion is still not public.