hack/stabilization-changes: Add --poll
Making this something that can run as a continuous robot.  If we
aren't ready for any promotions right now, poll again an hour or
whatever later to see if delays have elapsed or errata have been
published.  But don't notify Slack that frequently, because we'd drown
in distracting noise.

The idea is that the robot script would redeploy on a fresh build
after each promotion pull request lands, so we don't need to worry
about pulling to track master here.  There might be some races like:

1. Sync round notices 4.y.z's errata is public.
   a. Robot proposes 4.y.z into fast-4.y as #10
   b. Robot proposes 4.y.z into fast-4.(y-1) as #11
2. #10 passes CI and lands.  #11 goes into retesting.
3. New robot deploy tries to promote 4.y.z into fast-4.(y-1), but
   chokes on the branch-name collision.
4. #11 passes CI and lands.
5. New robot deploy is happy with the world.

But we can finess that branch-name collision handling in follow-up
work.