hack/stabilization-changes: Automatically follow upstream before new poll rounds
In single-shot mode, assume the caller has the local directory where
they want it to be.  For poll-mode, do that for the first round,
although there's a chance we've gone a bit stale since the image was
built (in which case we might fail while trying to recreate an
existing upstream branch, but no big deal).  But after waking up from
the poll-sleep for the next attempt, fetch and checkout the upstream
tip, to ensure we stay up to date.  At least with the graph-data; we
won't be picking up new Python code this way.

I'd somewhat prefer to trigger our BuildConfig with a GitHub webhook
to get this same effect (rebuilding based on the upstream tip, and
then bumping the deployment to run that new build) [1].  That would
get new Python picked up too.  But we can't do that conveniently
today, because we're deploying the cluster inside Red Hat's VPN, where
GitHub can't reach the cluster to poke the BuildConfig trigger.  One
benefit to this commits internal tracking is that we preserve the
next_notification time, so we don't flood the webhook channel with
waiting notifications.

While I'm at it, I'm also setting text=True [2], so the captured
output is a string, and I can drop the UTF-8 decoding.

I'm also adding clearer handling for "I'd like to create a promotion
branch, but that branch already exists locally".  This should generate
more readable messages when we're just waiting on presubmits/Tide to
merge an open pull request, vs. the current:

  FAILED channels/stable-4.6: Promote 4.6.38. It was promoted to the feeder fast-4.6 by aadc4bb8b0 (Merge pull request #920 from openshift/promote-4.6.38-to-fast-4.6, 2021-07-14) 2 days, 0:32:00.208424 ago. Command '['git', 'checkout', '-b', 'promote-4.6.38-to-stable-4.6', 'origin/master']' returned non-zero exit status 128.

[1]: https://docs.openshift.com/container-platform/4.7/cicd/builds/triggering-builds-build-hooks.html#builds-using-image-change-triggers_triggering-builds-build-hooks
[2]: https://docs.python.org/3/library/subprocess.html#subprocess.run