hack/stabilization-changes: Use the configured token to push
Avoid:

  Branch promote-4.6.39-to-stable-4.6 set up to track remote branch master from origin.
  [promote-4.6.39-to-stable-4.6 71a3666] channels/stable-4.6: Promote 4.6.39
   1 file changed, 1 insertion(+)
  fatal: could not read Username for 'https://github.com': No such device or address
  ERROR:   failed to promote 4.6.39 to stable-4.6: Command '['git', 'push', '-u', 'origin', 'promote-4.6.39-to-stable-4.6']' returned non-zero exit status 128.

I think that's Git saying "I need to authenticate this push.  Let me
ask for credentials using the TTY.  What!?  No TTY in this container!
Aaaah!".

I spent a while looking at [1], which sounded promising, but while
there is clear discussion on how to set up the cache helper and such,
and how to insert creds into the cache via TTY prompts, I couldn't
find a way to insert credentials without using a TTY.

So instead, this commit constructs a remote upstream URI, inserting
the configured GitHub token as user information [2].  And then we push
to that URI instead of pushing to the remote.  The token will be
visible to folks with /proc/ access, but they could already read it
out of our environment variables.  And we keep it off the local disk.

urlsplit returns a named tuple [3], and the _replace method allows
us to slot in our token user information [4].

The bulk of the new code is up before the branch checkout, so if we
raise an exception in there, we do it without poluting the local disk
state with an abandoned branch.

[1]: https://git-scm.com/docs/gitcredentials
[2]: https://datatracker.ietf.org/doc/html/rfc3986/#section-3.2.1
[3]: https://docs.python.org/3/library/urllib.parse.html#urllib.parse.urlsplit
[4]: https://docs.python.org/3/library/collections.html#collections.somenamedtuple._replace