Extend promotion delays
The most current minor goes to 96hr, and in order to avoid the previous
minor not having upgrade paths into the latest minor extend its delay to
reduce the likelihood of that happening. This may be moot until we actually
bring stable channel upgrades to 4.8 and perhaps by the time we achieve
that milestone we decide to get closer to 72, 48, 24 across 4.N, 4.N-1, 4.N-2.