channels/fast-4.8: Promote 4.8.9
4.8.9 was accidentally cut without an errata URI; compare:

  $ oc adm release info -o json quay.io/openshift-release-dev/ocp-release:4.8.9-x86_64 | jq .metadata
  {
    "kind": "cincinnati-metadata-v0",
    "version": "4.8.9",
    "previous": [
      "4.7.21",
      ...
      "4.8.7"
    ]
  }

and:

  $ oc adm release info -o json quay.io/openshift-release-dev/ocp-release:4.8.5-x86_64 | jq .metadata
  {
    "kind": "cincinnati-metadata-v0",
    "version": "4.8.5",
    "previous": [
      "4.7.21",
      ...
      "4.8.4"
    ],
    "metadata": {
      "description": "",
      "url": "https://access.redhat.com/errata/RHSA-2021:3121"
    }
  }

But it did actually get an errata, and that errata just shipped [1],
so the release is supported and needs to go into the fast channels.
It doesn't get into fast-4.9 at the moment, because the 'filter' set
on that channel only allows 4.9 releases, but we may need a manual
promotion of 4.8.9 into fast-4.9 later when we soften that filter.

[1]: https://access.redhat.com/errata/RHBA-2021:3247