blocked-edges/4.7.4*: Targeted edge blocking and version 1.1.0
So we can explain why we're blocking the different edges [1] (the
promql -> PromQL type change is in flight with [2]).

The zz in the filename for the vSphere hostname block ensures that one
sorts last alphabetically, because it's the broadest block, and legacy
Cincinnati services will prefer the final regular expression they load
for a given 'to' target.

This is basically a second attempt at my earlier 39bc2fb686
(blocked-edges/4.7.4*: Targeted edge blocking and version 1.1.0,
2021-09-01, #1056), which ended up getting reverted in da1254a39f
(Revert "blocked-edges/4.7.4*: Targeted edge blocking and version
1.1.0", 2021-09-21, #1078) because the production Cincinnati was mad
about the 1.1.0 version string.  [3] taught Cincinnati to relax, and
now that's live (and we never shipped any versions that would be mad
about 1.1.0 to customers, the 4.6.0 Update Service operator went out
before [4]).

[1]: https://github.com/openshift/enhancements/pull/821
[2]: https://github.com/openshift/enhancements/pull/910
[3]: https://github.com/openshift/cincinnati/pull/538
[4]: https://github.com/openshift/cincinnati/pull/314