hack/stabilization-changes: Allow local changes without branching
When we have a GitHub token, we checkout a branch and create a pull
request for each promotion.  When we don't have a token, we used to
just log the changes that needed to happen.  With this commit, a
token-less call will now edit the local files to perform all
appropriate promotions.  That allows this script (which understands
about tombstones and filters and other promotion criteria) to be used
to backfill channels when we expand filters (e.g. allowing 4.(y-1) in
fast-4.y after 4.y GAs, or allowing 4.(y-1) in stable-4.y after we
feel comfortable with the fast behavior.  That should make things like
8da5782432 (4.8 GA -- Fill in 4.7.z into fast-4.8, 2021-07-26, #947)'s
accidental promotion of the tombstoned 4.7.10 and similar less likely.
More on that in 4c53c7b4bd (Blocking all the incoming edges to 4.7.10,
4.7.14, 4.7.15, 4.7.20, 2021-09-03, #1035).

Drop the old backfill script from 2a02325043 (hack/backfill: Add a
backfilling script, 2020-11-12, #549), now that this new functionality
makes it obsolete.