hack/show-edges: Add a new script to show historical edges
A good deal of the loading logic is resurrected from the graph-util
script that was removed in 4b00fd6caf (hack/graph-util: Drop the
obsolete Python script, 2020-12-07, #578).  I'm still hoping that the
Rust tool grows a way to spit out single-shot Cincinnati JSON for each
channel, but until we grow that, this Python gives a useful stopgap.
For example, the addition of 4.8.9 to stable-4.8:

  $ git --no-pager log --oneline --first-parent 4fbfb85..2ac2547
  2ac2547 (origin/master, origin/HEAD, master) Merge pull request #1042 from openshift/promote-4.8.9-to-stable-4.8
  1faa0fe Merge pull request #1041 from openshift/promote-4.7.28-to-stable-4.7

looks like:

  $ diff -u <(hack/show-edges.py --revision 4fbfb85 stable-4.8) <(hack/show-edges.py --revision 2ac2547 stable-4.8)
  --- /dev/fd/63  2021-09-05 20:04:48.151716441 -0700
  +++ /dev/fd/62  2021-09-05 20:04:48.152716441 -0700
  @@ -1,6 +1,10 @@
   4.8.2 -> 4.8.3
   4.8.2 -> 4.8.4
   4.8.2 -(blocked)-> 4.8.5
  +4.8.2 -(blocked)-> 4.8.9
   4.8.3 -> 4.8.4
   4.8.3 -(blocked)-> 4.8.5
  +4.8.3 -(blocked)-> 4.8.9
   4.8.4 -(blocked)-> 4.8.5
  +4.8.4 -(blocked)-> 4.8.9
  +4.8.5 -> 4.8.9

Confirming that the 4.8.9 blocks are working as expected.