Increase fast to stable delay to 7d
Our telemetry analysis works best when we have a certain minimum
number of cluster days and our current delays weren't long enough.
As the fleet grows, participation in the fast channel grows, or we
improve our analysis we will revisit these values.

Aligning them all to the same value also ensures that we're less
likely to push something into stable-4.6 which cannot upgrade to
stable-4.7. We should also consider other solutions to this problem
such as only adding versions which do not match the channels X.Y
after that version has upgrade edges off of that, ie: 4.6.100 would
be added to stable-4.6 per normal delay, however it would not be
added to stable-4.7 until there's an upgrade from 4.6.100 to 4.7.N.
However that depends on shifting patterns toward an admin changing
to the desired channel and then upgrading. We probably still have a
large porition of cluster admins that upgrade as far as they can
then change channels and that solution wouldn't improve things for
them.