blocked-edges/4.8.*: Block 4.7 to 4.8 on SDN egress IP instability
Impact statement is [1].  All of the existing 4.8 releases are
impacted.  No 4.7 releases are impacted.  This commit blocks 4.7 ->
4.8 for all existing fast-4.8 4.8.z.

I've also included 4.8.15, which is currently only in candidate-4.8,
but which has no tombstones yet, so it's likely to make it into
fast-4.8.  If we end up tombstoning 4.8.15 for some reason, we can
drop that edge block.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=2008987#c51