channels/candidate-4.*: Sort versions for 4.6 and later
Using the SemVer-ish logic from hack/stabilization-changes.py, so when
we add feeders for these in a few releases it is easier to confirm
that there are no significant differences.

Also limit casting to integers to the first three fields
(major.minor.patch), to avoid comparison errors between 'fc' strings
and 0 integers from the hotfix releases.