internal-channels/candidate: Add a version-agnostic candidate channel
I've seeded it with the versions and tombstones from the 4.6 and later
candidate channels.  4.5 and earlier are end-of-life and unlikely to
receive new releases, so I've left them out.  Then I've adjusted the
version-capped candidate channels to feed off the version-agnostic
channel.

That pulls some 4.(y-1) FCs and other pre-releases into the
candidate-4.y channel, but that's not terrible, because folks are free
to install a 4.(y-1) FC and get from there to a 4.y FC using candidate
channels already.  The only difference here is how far along you need
to get before you can select the candidate-4.y channel.

The version-capped channels still retain some tombstones, because at
the moment they're still feeding the version-capped fast channels.
Once the version-capped fast channels start pulling from a future
version-agnostic fast channel, we'll be able to drop the
version-capped tombstone entries.

I'm personally in favor of making this channel something that's served
from Cincinnati and avaialble for folks to choose as a cluster
channel.  But until we have consensus around that, put the
version-agnostic channel in a separate directory where Cincinnati
won't notice it.