A prerequsite for 4.9 to 4.10 upgrades has been identified
https://bugzilla.redhat.com/show_bug.cgi?id=2037665

While 4.9.18 includes logic to ensure that clusters cannot upgrade
to 4.10 without addressing the pre-requisite it introduces a minor
regression where it incorrectly blips through Upgradeable=False. As
such we will file a subsequent bump to a build with the complete
set of changes when that 4.9.z is tagged. Bumping now will avoid
potential known issues in upgrading between 4.9 and 4.10, so we'll
bump twice.