hack/release-feature-freeze: Script feature freeze
So folks can call a script instead of reading docs to prepare a new
minor release for feature candidate builds.  Like 288b5daa57
(hack/release-ga: Automate GA steps, 2022-03-10, #1613), but for the
earlier lifecycle event.

I've left of the final:

  $ git add build-suggestions/* channels/*
  $ git commit

but we could do that in the future if we wanted to control the commit
message.

As described in 654f0faae1 (channels/candidate-4.10: Create this
channel, 2021-12-10, #1281), we don't want to create fast-* or later
channels at this point, because we're still far from GA, and we don't
want folks running 4.y.z to see a fast-4.(y+1) channel, get all
excited, switch to that channel, and then not see any 4.(y+1).z
targets for weeks.