hack/release-stable-minor: Script minor-update stablization
Continuing in the spirit of 288b5daa57 (hack/release-ga: Automate
GA steps, 2022-03-10, #1613) and d45fac659e
(hack/release-feature-freeze: Script feature freeze, 2022-04-06, #1720).
These lifecycle events aren't frequent, and automating them decreases
the odds that we fumble an unfamiliar transition.

I've left off the final:

  $ git add channels/*
  $ git commit

but we could do that in the future if we wanted to control the commit
message.