blocked-edges/4.10.*parallel-ceph_fsync: Wrap in topk
Continuing to clean up the PromQL from 707ec48c76
(blocked-edges/4.10.*: Parallel ceph_fsync risk for 4.10.4 and later,
2022-05-18, #1915), which I'd previously revised in da76bd14ab
(blocked-edges/4.10.*parallel-ceph_fsync: Handle multiple
ceph_health_status series, 2022-05-20, #1932).  We'd still been
getting:

  invalid PromQL result length must be one, but is 2

The semantics for 'or' are [1]:

  vector1 or vector2 results in a vector that contains all original
  elements (label sets + values) of vector1 and additionally all
  elements of vector2 which do not have matching label sets in
  vector1.

I'd missed the "+ values" part of it, but it meant that when:

  label_replace(group(ceph_health_status), "ceph", "yes", "", "")

hit, returning a series like {ceph="yes"}=1, and:

  label_replace(0 * group(cluster_version), "ceph", "no", "", "")

hit, returning a series like {ceph="no"}=0, we'd return both of those
series.  This commit wraps the expression in a topk(1, ...) [2], so we
only get the largest series (i.e. {ceph="yes"}=1 wins).

[1]: https://prometheus.io/docs/prometheus/latest/querying/operators/#logical-set-binary-operators
[2]: https://prometheus.io/docs/prometheus/latest/querying/operators/#aggregation-operators