hack/stabilization-changes: Gate 4.(y-1).z into fast-4.y, etc. on update paths to 4.y
We sometimes tombstone or otherwise do not release a 4.y.z in a given
week, when we do release 4.(y-1).z'. Hypothetical example:

1. Some bug shows up which makes 4.98.0-to-4.99 conditionally risky,
   and we've raised minor_min to 4.98.25 for new 4.99 builds.
2. Week 1: 4.98.90 and 4.99.50 are built. All is well.
3. Week 2: 4.98.91 and 4.99.51 are built, but 4.99.51 gets tombstoned.
4. Week 3: 4.98.91 enters fast channels, including fast-4.99.
5. Alice is on 4.98.0, and decides to move to 4.99 via
   fast-4.99. There are no recommended updates to 4.99.50 or 4.9.51
   because of the long-standing minor_min floor.
6. Alice sees recommendations to 4.98.90 and 4.98.91, and heads off to
   4.98.91 because it fixes more bugs.
7. Oh no, there are no recommendations to get from 4.98.91 to
   4.99. Alice is stuck.
8. Week 4: 4.99.52 enters fast channels, with an update path from
   4.98.91, and Alice can finally update.

What we want is to withhold 4.98.91 from fast-4.99 until it has an
update path into 4.99, so when Alice tries to get to 4.99, we send her
through 4.98.90, saving her the week+ of waiting for a new 4.99.z
release that can update from 4.98.91.

Most of the difficulty here is figuring out whether there will be an
update path.  We could teach the script to compute those based on
local graph-data information and registry requests like
hack/show-edges.py does.  But walking Quay tags and pulling and
inspecting release images is slow and tedious, and we probably don't
need this sort of fanciness for candidate-... channels. So this commit
sticks with our current behavior for candidate-..., uses the script's
existing update service request logic to get update recommendations
for candidate-4.y, and expects those same recommendations to hold true
for fast-4.y and later if the source and target releases were both in
the fast-4.y or later channel (which is how Cincinnati works today,
and likely to be how it continues to work in the future).

I'm looking at multiple hops.  For example, with 4.(y-2).z,
4.(y-1).z', and 4.y.z" releases, 4.(y-1).z' will only enter fast-4.10
if it has a path to 4.y.z".  But by the time we go to consider
4.(y-2).z, the previously-confirmed 4.(y-1).z'-to-4.y.z" update may no
longer be recommended, and in that case we'd want to hold 4.(y-2).z
out of the channel until 4.(y-1).z'-to-4.y.z" was restored or an
alternate path to 4.y became unconditionally recommended.